pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xai-investment-banking"
version = "0.1.0"
description = "Dissertation project: risk scoring, valuation, and explainability for investment banking."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy",
  "pandas",
  "requests",
  "python-dotenv",
  "tqdm",
  "pyyaml",
  "scipy",
  "matplotlib",
  "shap",
]

[tool.setuptools]
packages = ["src"]
